+++ src/lz-ui/src/components.rs
@@ -1,0 +1,5 @@
+mod bookmarks;
+pub use bookmarks::*;
+
+mod bookmark;
+pub use bookmark::*;
